requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
influxdb-client==1.38.0
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
APScheduler==3.10.4
pytest==7.4.3
httpx==0.25.2
aiofiles==23.2.1
